refacto(store): implementing the component StateHandler
